Convert to new style of defaults/pillar lookups
Convert to the more modern style of merging defaults and pillar lookups all in
the map.jinja file.

This is a stepping stone to some more changes I have locally. So if it looks
like incomplete, that's probably because it is. This should work the same way
as before though.